@@ -7,7 +7,7 @@ requires = ["hatchling"]
 name = "sincei"
 description = "A user-friendly toolkit for QC, counting, clustering and plotting of single-cell (epi)genomics data "
 readme = "README.md"
-requires-python = ">=3.8"
+requires-python = ">=3.12"
 dynamic = ["version"]
 license = {file = "LICENCE.txt"}
 authors = [
@@ -40,7 +40,8 @@ dependencies = [
     "session-info",
     "torch",
     "mctorch-lib",
-    "gensim"
+    "gensim",
+    "ruptures",
     ]
 
 [project.scripts]
@@ -49,6 +50,7 @@ scClusterCells = "sincei.scClusterCells:main"
 scCombineCounts = "sincei.scCombineCounts:main"
 scCountQC = "sincei.scCountQC:main"
 scCountReads = "sincei.scCountReads:main"
+scFindVCRs = "sincei.scFindVCRs:main"
 scFilterBarcodes = "sincei.scFilterBarcodes:main"
 scFilterStats = "sincei.scFilterStats:main"
 scJSD = "sincei.scJSD:main"
@@ -60,13 +62,13 @@ dev = [
     # CLI for bumping the version number
     "bump2version",
     "pre-commit",
-    "twine>=4.0.2"
+    "twine>=6.2.0"
 ]
 doc = [
-    "sphinx>=8",
-    "sphinx-book-theme>=0.3.3",
+    "sphinx>=9.1.0",
+    "sphinx-book-theme>=1.1.4",
     "myst-nb",
-    "sphinxcontrib-bibtex>=1.0.0",
+    "sphinxcontrib-bibtex>=2.6.5",
     "sphinx-autodoc-typehints",
     "readthedocs-sphinx-ext",
     "sphinx-argparse",
@@ -107,7 +109,7 @@ skip_glob = ["docs/*"]
 
 [tool.black]
 line-length = 120
-target-version = ['py310']
+target-version = ['py312']
 include = '\.pyi?$'
 exclude = '''
 (
