@@ -14,7 +14,6 @@ import re
 import sys
 from importlib import metadata
 
-
 # -- Project information -----------------------------------------------------
 
 project = "sincei"
@@ -73,6 +72,7 @@ exclude_patterns = [
 #     "numpy", "pandas", "scipy", "scanpy", "anndata", "mudata", "umap", "leidenalg",
 #     "matplotlib", "networkx", "igraph", "torch", "mctorch", "deeptools", "gensim",
 #     "tqdm", "deeptoolsintervals", "sklearn", "pysam", "pyBigWig", "joblib", "py2bit",
+#     "ruptures",
 # ]
 
 
