@@ -1,1 +1,1 @@
-__version__ = "0.5.2"
+__version__ = "0.6.0"
