@@ -289,12 +289,8 @@ def main(args=None):
         checkBAMtag(x, bam, args.cellTag)
         if args.groupTag:
             checkBAMtag(x, bam, args.groupTag)
-            sys.stderr.write(
-                "--groupTag is not implemented for scFilterStats yet! \
-            Please split your BAM file by {} and re-run scFilterStats. \n".format(
-                    args.groupTag
-                )
-            )
+            sys.stderr.write("--groupTag is not implemented for scFilterStats yet! \
+            Please split your BAM file by {} and re-run scFilterStats. \n".format(args.groupTag))
             exit(1)
         x.close()
 
