@@ -3,37 +3,32 @@
 
 import argparse
 import sys
-
-from sincei._version import __version__
+from importlib import metadata
 
 
 def parse_arguments(args=None):
     parser = argparse.ArgumentParser(
         formatter_class=argparse.RawDescriptionHelpFormatter,
-        description="""
+        description=f"""
 sincei is a suite of command-line tools developed for a user-friendly analysis of single-cell sequencing data.
-Version: {}
+Version: {metadata.version("sincei")}
 
 Each tool begins with the prefix sc<tool_name>, such as:
 
  $ scBulkCoverage -b file1.bam -g groupinfo.txt -o coverage
 
-[ Tools for a typical single-cell analysis workflow ] (WIP: work in progress/not available yet)
+[ Tools for a typical single-cell analysis workflow ]
 
     scFilterBarcodes        Identify and filter cell barcodes from BAM file (for droplet-based single-cell seq)
     scFilterStats           Produce per-cell statistics after filtering reads by user-defined criteria.
+    scFindVCRs              Call variable chromatin regions (VCRs) from binned chromatin data.
+    scScoreFeatures         Calculate gene activity scores from chromatin features/bins.
     scCountReads            Counts reads for each barcode on genomic bins or user-defined features.
     scCountQC               Perform quality control and filter the output of scCountReads.
-    scCombineCounts         [WIP] Concatenate/merge the counts from different samples/batches or modalities
+    scCombineCounts         Concatenate/merge the counts from different samples/batches or modalities.
     scClusterCells          Perform dimensionality reduction and clustering on the output of scCountReads.
     scBulkCoverage          Get pseudo-bulk coverage per group using a user-supplied cell->group mapping (output of scClusterCells).
-    scBAMops                Modify a BAM file to group cells (using cell barcodes), or filter/shift mapped reads.
-    scFindMarkers           [WIP] Find marker genes per group, given the output of scCountReads and a user-defined group.
-    scFeaturePlot           [WIP] Plot the counts for a given feature on a UMAP or on a (IGV-style) genomic-track.
-
-""".format(
-            __version__
-        ),
+""",
     )
 
     return parser
